+++ python/GafferCortexUITest/CompoundPlugValueWidgetTest.py
@@ -61,26 +61,5 @@ class CompoundPlugValueWidgetTest( GafferUITest.TestCase ) :
 		self.assertIsInstance( pw.childPlugValueWidget( n["c"]["i"] ), GafferUI.PlugValueWidget )
 		self.assertIsInstance( pw.childPlugValueWidget( n["c"]["s"] ), GafferUI.PlugValueWidget )
 
-	def testChildReadOnlyStatus( self ) :
-
-		n = Gaffer.Node()
-		n["c"] = Gaffer.Plug()
-		n["c"]["i"] = Gaffer.IntPlug()
-		n["c"]["s"] = Gaffer.StringPlug()
-
-		w = CompoundPlugValueWidget( n["c"] )
-		w.setReadOnly( True )
-
-		iw = w.childPlugValueWidget( n["c"]["i"] )
-		sw = w.childPlugValueWidget( n["c"]["s"] )
-
-		self.assertEqual( iw.getReadOnly(), True )
-		self.assertEqual( sw.getReadOnly(), True )
-
-		w.setReadOnly( False )
-
-		self.assertEqual( iw.getReadOnly(), False )
-		self.assertEqual( sw.getReadOnly(), False )
-
 if __name__ == "__main__":
 	unittest.main()