+++ python/GafferUI/CompoundDataPlugValueWidget.py
@@ -83,15 +83,6 @@ class CompoundDataPlugValueWidget( GafferUI.PlugValueWidget ) :
 		self.__layout = GafferUI.PlugLayout( plug )
 		self.__column[0] = self.__layout
 
-	def setReadOnly( self, readOnly ) :
-
-		if readOnly == self.getReadOnly() :
-			return
-
-		GafferUI.PlugValueWidget.setReadOnly( self, readOnly )
-
-		self.__layout.setReadOnly( readOnly )
-
 	def childPlugValueWidget( self, childPlug ) :
 
 		return self.__layout.plugValueWidget( childPlug )