+++ python/GafferUI/NodeUI.py
@@ -111,7 +111,6 @@ class NodeUI( GafferUI.Widget ) :
 		GafferUI.Widget.__init__( self, topLevelWidget, **kw )
 
 		self.__node = node
-		self.__readOnly = False
 
 	## Returns the node the ui represents.
 	def node( self ) :
@@ -125,17 +124,6 @@ class NodeUI( GafferUI.Widget ) :
 	def plugValueWidget( self, plug ) :
 
 		return None
-
-	## \deprecated
-	def setReadOnly( self, readOnly ) :
-
-		assert( isinstance( readOnly, bool ) )
-		self.__readOnly = readOnly
-
-	## \deprecated
-	def getReadOnly( self ) :
-
-		return self.__readOnly
 
 	## Creates a NodeUI instance for the specified node.
 	@classmethod
@@ -166,9 +154,7 @@ class NodeUI( GafferUI.Widget ) :
 	@staticmethod
 	def appendPlugDeletionMenuDefinitions( plugOrPlugValueWidget, menuDefinition ) :
 
-		readOnlyUI = False
 		if isinstance( plugOrPlugValueWidget, GafferUI.PlugValueWidget ) :
-			readOnlyUI = plugOrPlugValueWidget.getReadOnly()
 			plug = plugOrPlugValueWidget.getPlug()
 		else :
 			plug = plugOrPlugValueWidget
@@ -184,7 +170,7 @@ class NodeUI( GafferUI.Widget ) :
 		if len( menuDefinition.items() ) :
 			menuDefinition.append( "/DeleteDivider", { "divider" : True } )
 
-		menuDefinition.append( "/Delete", { "command" : functools.partial( NodeUI.__deletePlug, plug ), "active" : not readOnlyUI and not Gaffer.MetadataAlgo.readOnly( plug ) } )
+		menuDefinition.append( "/Delete", { "command" : functools.partial( NodeUI.__deletePlug, plug ), "active" : not Gaffer.MetadataAlgo.readOnly( plug ) } )
 
 	@staticmethod
 	def __deletePlug( plug ) :