+++ Changes.md
@@ -55,6 +55,11 @@ Breaking Changes
 - CatalogueUI : Hid OutputIndexColumn from public API.
 - ContextAlgo : Removed use of the `ui:scene:expandedPaths` context variable. Any code directly accessing `ui:scene:expandedPaths` should instead use the `getExpandedPaths()/setExpandedPaths()/expand()/expandDescendants()` methods provided by `ContextAlgo`.
 - SceneGadget : Removed `setExpandedPaths()` and `getExpandedPaths()` methods. `setVisibleSet()` and `getVisibleSet()` are now used instead.
+- GafferUI : Removed the `setReadOnly()` and `getReadOnly()` methods from the following widgets :
+  - `GafferUI.PlugValueWidget`
+  - `GafferUI.NodeEditor`
+  - `GafferUI.NodeUI`
+  - `GafferUI.PlugLayout`
 
 1.1.x.x (relative to 1.1.4.0)
 =======